+++ src/aiqclib/train/step2_validate_model/kfold_validation_suite.py
@@ -143,6 +143,7 @@ class KFoldValidationSuite(ValidationBase):
                 current_fold_model.build()
                 self.models[comp_key].append(current_fold_model)
 
+                current_fold_model.target_name = target_name
                 current_fold_model.test_set = (
                     self.training_sets[target_name]
                     .filter(pl.col("k_fold") == (k + 1))